+++ src/lib.rs
@@ -45,7 +45,7 @@ pub use args::{EnvVar, parse_argv_envp};
 pub use console::Console;
 pub use nix_bytes::NixBytes;
 pub use nix_str::NixString;
-pub use print::{__print_err, __print_str};
+pub use print::{__format, __print_err, __print_str};
 pub use syscall::{Errno, SyscallArg, SyscallNum};
 pub(crate) use syscall::{syscall, syscall_result};
 pub use test_framework::custom_test_runner;