+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `clear`: Clears the terminal screen.
 - `File::read_to_bytes`: Reads the whole file into a byte vector.
+- `format!`: Just like the `std` macro!
 
 ### Changed